refactor: split frontend index.ts into focused modules
Closes #155 (step 1 — mechanical split, zero behaviour changes).

Split the 3466-line index.ts into focused modules following dependency order:

  types.ts        — all TypeScript types/interfaces
  translations.ts — translation system (cache, fetch, localize)
  helpers.ts      — utility & domain helpers (formatDate, favorites, etc.)
  ui.ts           — render helpers (renderCardHeader, renderPermitSelect, etc.)
  base.ts         — BaseLocalizedCard, BaseCardEditor, registration helpers
  editor-parking.ts  — CityVisitorParkingCardEditor + getCardConfigForm
  editor-active.ts   — CityVisitorParkingActiveCardEditor + getActiveCardConfigForm
  card-parking.ts    — CityVisitorParkingNewReservationCard
  card-active.ts     — CityVisitorParkingActiveCard
  index.ts        — entry point (imports only, triggers side-effects)

Build passes, tsc --noEmit clean, ESLint + Prettier pass.
